refactor: rename railties package to trailties
Renames the `@blazetrails/railties` package to `@blazetrails/trailties` to
match the `trails` project name. Updates all references across configs,
compare scripts, website vite/vitest aliases, and docs. Rails source paths
inside fetch scripts are preserved since they point at the upstream Rails
`railties/` directory.

Also renames the `actionpack/src/actioncontroller/railties/` subfolder to
`trailties/` for consistency.
