+++ README.md
@@ -25,7 +25,7 @@ The goal is to be **100% API compatible with Rails**, matching behavior **test f
 | `@blazetrails/actionpack` | [ActionController](https://api.rubyonrails.org/classes/ActionController.html) | **67.6%** | **28.3%** | Controller layer, rendering, filters, parameters       |
 |                           | [ActionDispatch](https://api.rubyonrails.org/classes/ActionDispatch.html)     | **6.1%**  | **37.3%** | Routing, middleware stack, cookies, sessions, security |
 | `@blazetrails/actionview` | [ActionView](https://api.rubyonrails.org/classes/ActionView.html)             | **3.7%**  | **5.1%**  | Templates, rendering, view helpers                     |
-| `@blazetrails/railties`   | [Railties](https://api.rubyonrails.org/classes/Rails.html)                    | **0.1%**  | **3.9%**  | CLI, generators, application bootstrap                 |
+| `@blazetrails/trailties`  | [Railties](https://api.rubyonrails.org/classes/Rails.html)                    | **0.1%**  | **3.9%**  | CLI, generators, application bootstrap                 |
 
 **Tests** = `test:compare` — matches our test names against the Rails test suite. **API** = `api:compare` — matches individual public methods against Rails source (method-level, not class/module wrappers). Rack doesn't have API comparison yet (it's not a Rails gem).
 
@@ -135,7 +135,7 @@ packages/
   rack/             — Web server interface, middleware, request/response
   actionpack/       — ActionDispatch (routing, cookies, sessions) and ActionController
   actionview/       — Templates, rendering, view helpers
-  railties/         — CLI, generators, application bootstrap
+  trailties/        — CLI, generators, application bootstrap
 ```
 
 ## Disclaimer
