+++ docs/launch-roadmap.md
@@ -12,7 +12,7 @@ rewritten in TypeScript" — backed by a working demo and rock-solid core packag
 | actiondispatch   | 406/1,432         | 28.4%    | In progress (27 misplaced) |
 | actioncontroller | 7/1,860           | 0.4%     | Early (229 misplaced)      |
 | actionview       | 0/2,497           | 0%       | Not started                |
-| railties         | 93/2,411          | 3.9%     | Early                      |
+| trailties        | 93/2,411          | 3.9%     | Early                      |
 | **Overall**      | **10,296/21,890** | **47%**  |                            |
 
 Overall includes completed packages (arel 100%, activemodel 100%, rack 100%).
@@ -52,7 +52,7 @@ a model, run migrations, start a server, hit JSON and HTML endpoints.
 
 ### Nice to have
 
-- [ ] **railties** — application bootstrap, configuration. Would be great for
+- [ ] **trailties** — application bootstrap, configuration. Would be great for
       `trails new` to produce a real runnable app, but not blocking.
 - [ ] Marketing site polish (part of Frontiers split PR 1)
 - [ ] README overhaul with code examples and badges
