+++ packages/trailties/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@blazetrails/railties",
+  "name": "@blazetrails/trailties",
   "version": "0.1.0",
   "type": "module",
   "bin": {
