+++ packages/trailties/src/bin.ts
