+++ packages/trailties/src/commands/destroy.test.ts
