+++ packages/trailties/src/commands/destroy.ts
