+++ packages/trailties/src/commands/generate.test.ts
