+++ packages/trailties/src/commands/generate.ts
