+++ packages/trailties/src/commands/new.test.ts
