+++ packages/trailties/src/commands/new.ts
