+++ packages/trailties/src/commands/routes.ts
