+++ packages/trailties/src/generators/controller-generator.ts
