+++ packages/trailties/src/generators/model-generator.ts
