+++ packages/trailties/src/generators/scaffold-generator.test.ts
