+++ packages/trailties/src/generators/scaffold-generator.ts
