+++ packages/trailties/src/server/vite-plugin.ts
