+++ tsconfig.json
@@ -25,7 +25,7 @@
     { "path": "packages/rack" },
     { "path": "packages/actionview" },
     { "path": "packages/actionpack" },
-    { "path": "packages/railties" }
+    { "path": "packages/trailties" }
   ],
   "exclude": ["node_modules", "dist"]
 }
